+++ osmnx/io.py
@@ -421,6 +421,15 @@ def _convert_node_attr_types(G, dtypes=None):
     G : networkx.MultiDiGraph
     """
     for _, data in G.nodes(data=True):
+        # first, eval stringified lists, dicts, or sets to convert them to objects
+        # lists, dicts, or sets would be custom attribute types added by a user
+        for attr, value in data.items():
+            if (value.startswith("[") and value.endswith("]")) or (
+                value.startswith("{") and value.endswith("}")
+            ):
+                with contextlib.suppress(SyntaxError, ValueError):
+                    data[attr] = ast.literal_eval(value)
+
         for attr in data.keys() & dtypes.keys():
             data[attr] = dtypes[attr](data[attr])
     return G
@@ -446,10 +455,13 @@ def _convert_edge_attr_types(G, dtypes=None):
         # remove extraneous "id" attribute added by graphml saving
         data.pop("id", None)
 
-        # first, eval stringified lists to convert them to list objects
+        # first, eval stringified lists, dicts, or sets to convert them to objects
         # edge attributes might have a single value, or a list if simplified
+        # dicts or sets would be custom attribute types added by a user
         for attr, value in data.items():
-            if value.startswith("[") and value.endswith("]"):
+            if (value.startswith("[") and value.endswith("]")) or (
+                value.startswith("{") and value.endswith("}")
+            ):
                 with contextlib.suppress(SyntaxError, ValueError):
                     data[attr] = ast.literal_eval(value)